requirements.txt
selenium~=4.34.2
pillow~=11.3.0
requests~=2.32.4
python-dotenv~=1.1.1